reorganized workspace for dev clarity
